REFACTOR: User Entity 로직 변화에 따른 Review Entity 수정
