+++ src/test/java/io/vtou/vitamintou/domain/review/domain/enums/ReviewScoreTest.java
@@ -3,7 +3,7 @@ package io.vtou.vitamintou.domain.review.domain.enums;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import io.vtou.vitamintou.domain.review.domain.exception.ReviewException;
+import io.vtou.vitamintou.domain.review.exception.ReviewException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
